Added script\script-development.py
**Note**

The default framework for a python application,
contains a go-to script-development.py,
as well as a go-to script-production.py.

The file script-production.py is intended,
as an environment for the creation of a,
complete product. &

The file script-development.py is,
intended as a development environment,
for an application or potential product.

As you probably already guessed.